Move resource models to core
