+++ common/lessons-data/src/main/java/app/ss/lessons/data/repository/quarterly/QuarterliesRepositoryImpl.kt
@@ -26,7 +26,7 @@ import com.cryart.sabbathschool.core.extensions.prefs.SSPrefs
 import com.cryart.sabbathschool.core.misc.SSConstants
 import app.ss.lessons.data.model.Language
 import app.ss.lessons.data.model.SSQuarterly
-import app.ss.lessons.data.response.Resource
+import com.cryart.sabbathschool.core.response.Resource
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.FirebaseDatabase
