+++ features/lessons/src/test/java/com/cryart/sabbathschool/lessons/ui/quarterlies/QuarterliesViewModelTest.kt
@@ -25,7 +25,7 @@ package com.cryart.sabbathschool.lessons.ui.quarterlies
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import app.ss.lessons.data.model.SSQuarterly
 import app.ss.lessons.data.repository.quarterly.QuarterliesRepository
-import app.ss.lessons.data.response.Resource
+import com.cryart.sabbathschool.core.response.Resource
 import com.cryart.sabbathschool.core.extensions.arch.observeFuture
 import com.cryart.sabbathschool.core.extensions.prefs.SSPrefs
 import com.cryart.sabbathschool.core.model.ViewState
