+++ features/readings/src/test/java/com/cryart/sabbathschool/readings/ReadingViewModelTest.kt
@@ -5,8 +5,9 @@ import app.ss.lessons.data.model.SSDay
 import app.ss.lessons.data.model.SSLesson
 import app.ss.lessons.data.model.SSLessonInfo
 import app.ss.lessons.data.repository.lessons.LessonsRepository
-import app.ss.lessons.data.response.Resource
+import com.cryart.sabbathschool.core.response.Resource
 import com.cryart.sabbathschool.core.misc.SSConstants
+import com.cryart.sabbathschool.core.model.UiState
 import com.cryart.sabbathschool.readings.components.model.AppBarData
 import com.cryart.sabbathschool.readings.components.model.ErrorData
 import com.cryart.sabbathschool.readings.components.model.ReadingDaysData
